server/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
autopep8==1.5.5
boto3==1.17.12
botocore==1.20.12
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.4
mypy-extensions==0.4.3
numpy==1.20.1
pandas==1.2.2
pathspec==0.8.1
Pillow==8.1.0
pycodestyle==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
regex==2020.11.13
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
